requirements.txt
selenium
typer
beautifulsoup4
python-dotenv
